Replace massive libsodium subtree with a submodule
Forked to ambiata/libsodium and created an `ambiata` branch from the
tag 1.0.11, then applied the autoconf patch from 547d0e1.

The reason I used a subtree initially was I found with `scrypt` that
it made it simpler to review/patch library code, but I have been
convinced that the negatives outweigh the positives in this case
(`libsodium` is massive, we really want to avoid patching it if we can
possibly avoid it, and it's consistent with how we solve this problem
elsewhere).